standalone/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
python-dotenv
typing-extensions
uuid
pyyaml
passlib[bcrypt]
python-jose[cryptography]
pydantic[email]
geopy
ortools
osmnx
networkx
osmium
scikit-learn
httpx